requirements.txt
deep_translator==1.11.4
docx==0.2.4
PyPDF2==3.0.1
python_docx==0.8.11
python_pptx==0.6.23
streamlit==1.23.1
